REALESTATE/requirements.txt
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
openai==0.27.8
playwright==1.43.0
APScheduler==3.10.1.post1
pytz==2023.3
sqlalchemy==1.4.49
folium==0.14.0
shapely==2.0.1
geopandas==0.13.2
python-dateutil==2.8.2
tqdm==4.65.0
